GoodProject/Orders/requirements/prod.txt
fastapi
request
uvicorn[standard]